rework logic to use dictionary based item spawning, selling and transport and have a central building manager that fires commands rather than the items themselves
buildings now inherrit from base building class with setup method
buildings now use building manager for certain logic
items no longer call logic but recive orders instead

+++ Assets/Scripts/Managers/CurrencyManager.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public class CurrencyManager : SingletonBase<CurrencyManager>
 {
-    public int currentCurrencyValue;
+    public float currentCurrencyValue;
     // Start is called once before the first execution of Update after the MonoBehaviour is created
     void Start()
     {
@@ -16,13 +16,13 @@ public class CurrencyManager : SingletonBase<CurrencyManager>
         
     }
 
-    public void AddCurrency(int currencyValue)
+    public void AddCurrency(float currencyValue)
     {
         currentCurrencyValue += currencyValue;
         UiManager.Instance.UpdateCurrency(currentCurrencyValue);
     }
 
-    public void RemoveCurrency(int currencyValue)
+    public void RemoveCurrency(float currencyValue)
     {
         currentCurrencyValue -= currencyValue;
         UiManager.Instance.UpdateCurrency(currentCurrencyValue);
